Normalized indentation (tabs)
The majority of this project uses single tabs for indentation, but there are a
lot of places where spaces are used. It's not really that important which you
use, but it is important that a project is internally consistent (for instance,
besides the philosophical annoyance, some users may have tabs set to a size
other than four spaces, which makes the inconsistency very jarring). As such, I
have converted the entire project to use tabs.

For anyone curious, you can set Xcode 4's indentation style for a specific
project by selecting the project in the source pane on the left, and opening the
File Inspector in the right pane, and choosing Tabs under Text Settings =>
Indent Using.