Fixed content view tile size for Retina iPad
The previous method would set a fixed tile size of 1024x1024. This value is cropped by the clip rect of the tiled layer to the actual size but without multiplying it by scale. This caused the partial drawing method in layout frame to break down on Retina iPad.

The new calculation gets the larger dimension and multiplies it with the screen scale to always get tiles that cover the entire width of the display regardless of orientation.

Fixes #145