Fixed problem with images contained in a href.
The merging of the href would cause the images to be merged as well. Now the creation of the link GUIDs happens when the link property of a HTML element is set and will be transferred to all child elements. This way a href that consists of text, image, text still gets the same link GUID.