Prevent justification of the final lines of paragraphs
Justify is still justifying the last line of some paragraphs, despite
pull request #180/ issue #179.

There are two bugs. First, the calculation of isAtEndOfParagraph on
line 172 of DTCoreTextLayoutFrame.m is wrong, checking to see if the
line starts at the end of the paragraph. Second, a br doesn't end the
paragraph but should suppress justification as if it was the end of the
paragraph.

Both issues are solved with this change.