improved handling of margin and padding CSS styles
margins are presently only considered for block-level elements. they are added for left and right. bottom margin becomes the space after a block = paragraphSpacing.

There is a workaround in place because the left padding interferes with list indenting

What should really be done here is calculate the cumulative block and/or list indenting from the margin values of the parents