Load available fonts into override table
Now the +initialize of DTCoreTextFontDescriptor gets a list of all available fonts and registers them in the override table. This way fonts can be found even if they are not registered in the override plist, in fact the override plist can be reduced or omitted.

This fixes a problem on iOS 5 where some fonts could not be matched if their family name contained a space.

Fixes #365