Refactored relayout after lazy image loading into DTAttributedTextView.
Without an extra setNeedsLayout an lazy image would not show up since we switched to a tiled layer for the content view and inside scrollviews the scroll view is responsible for the lay outing of just the visible custom subviews.

This reduces the code to put into lazyImageView:didChangeImageSize:. I also change the sizing behavior for zero size attachments. They still request a subview now.

Fixes #368