Removed UIKit calls from background thread
When loading a file url, there were constant issues with UIKit methods being called in the background. The issue was that -[DTLazyImageView loadImageAtURL:] always runs itself in the background. The networking code surly doesn't cause any performance concern, however, loading the file from disk might. So instead of running the entire method in the background, we only move to the background in cases where it would be beneficial, and even then, move back to the main thread to update everything.