Added support for Cocoapods modules
In order to be compiled as a module framework, we cannot reference anything in our headers that are not included in the module, or referenced as a separate framework (ie #import <DTFoundation/DTSomething.h>). If we simply converted our #imports to frameworks imports, it breaks the sample app, and probably other apps that include the framework in a similar way.

However, if we move a single protocol reference to the .m file instead of the header, the only DTFoundation imports left is DTWeakSupport.h. This file is used to support iOS 4.3 which doesn't have weak variables. If we change the minimum version to iOS 5, we can remove our dependency on that file completely. Unfortunately, that means we break support for iOS 4.3, but it has been 4 years now.